Add multi-arch architecture

+++ integration/global_test.go
@@ -43,10 +43,18 @@ func testGlobal(t *testing.T, context spec.G, it spec.S) {
 		})
 
 		it.After(func() {
-			Expect(docker.Container.Remove.Execute(container.ID)).To(Succeed())
-			Expect(docker.Image.Remove.Execute(image.ID)).To(Succeed())
-			Expect(docker.Volume.Remove.Execute(occam.CacheVolumeNames(name))).To(Succeed())
-			Expect(os.RemoveAll(source)).To(Succeed())
+			if container.ID != "" {
+				Expect(docker.Container.Remove.Execute(container.ID)).To(Succeed())
+			}
+			if image.ID != "" {
+				Expect(docker.Image.Remove.Execute(image.ID)).To(Succeed())
+			}
+			if name != "" {
+				Expect(docker.Volume.Remove.Execute(occam.CacheVolumeNames(name))).To(Succeed())
+			}
+			if source != "" {
+				Expect(os.RemoveAll(source)).To(Succeed())
+			}
 		})
 
 		it("builds and runs", func() {
@@ -60,14 +68,14 @@ func testGlobal(t *testing.T, context spec.G, it spec.S) {
 				WithPullPolicy("never").
 				WithBuildpacks(buildpacksArray...).
 				WithEnv(map[string]string{
-					"BP_COMPOSER_INSTALL_GLOBAL": "friendsofphp/php-cs-fixer",
+					"BP_COMPOSER_INSTALL_GLOBAL": "friendsofphp/php-cs-fixer=^3.40.0",
 					"BP_LOG_LEVEL":               "DEBUG",
 					"BP_PHP_SERVER":              "nginx",
 				}).
 				Execute(name, source)
 			Expect(err).ToNot(HaveOccurred(), logs.String)
 
-			Expect(logs).To(ContainSubstring("Running 'composer global require --no-progress friendsofphp/php-cs-fixer'"))
+			Expect(logs).To(ContainSubstring("Running 'composer global require --no-progress friendsofphp/php-cs-fixer=^3.40.0'"))
 
 			container, err = docker.Container.Run.
 				WithEnv(map[string]string{"PORT": "8765"}).
@@ -89,7 +97,7 @@ func testGlobal(t *testing.T, context spec.G, it spec.S) {
 				WithPullPolicy("never").
 				WithBuildpacks(buildpacksArray...).
 				WithEnv(map[string]string{
-					"BP_COMPOSER_INSTALL_GLOBAL": "friendsofphp/php-cs-fixer",
+					"BP_COMPOSER_INSTALL_GLOBAL": "friendsofphp/php-cs-fixer=^3.40.0",
 					"BP_LOG_LEVEL":               "DEBUG",
 					"BP_PHP_SERVER":              "nginx",
 				}).
@@ -102,7 +110,7 @@ func testGlobal(t *testing.T, context spec.G, it spec.S) {
 				WithPullPolicy("never").
 				WithBuildpacks(buildpacksArray...).
 				WithEnv(map[string]string{
-					"BP_COMPOSER_INSTALL_GLOBAL": "friendsofphp/php-cs-fixer",
+					"BP_COMPOSER_INSTALL_GLOBAL": "friendsofphp/php-cs-fixer=^3.40.0",
 					"BP_LOG_LEVEL":               "DEBUG",
 					"BP_PHP_SERVER":              "nginx",
 				}).
@@ -112,7 +120,7 @@ func testGlobal(t *testing.T, context spec.G, it spec.S) {
 
 			Expect(firstID).To(Equal(image.ID))
 
-			Expect(logs).To(ContainSubstring("Running 'composer global require --no-progress friendsofphp/php-cs-fixer'"))
+			Expect(logs).To(ContainSubstring("Running 'composer global require --no-progress friendsofphp/php-cs-fixer=^3.40.0'"))
 
 			container, err = docker.Container.Run.
 				WithEnv(map[string]string{"PORT": "8765"}).
